requirements.txt
numpy
matplotlib
scikit-learn
requests
pillow
python3-tk
tensorflow
